Refactor

+++ src/main/java/shpp/azaika/dao/CategoryDAO.java
@@ -3,7 +3,6 @@ package shpp.azaika.dao;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import shpp.azaika.dto.CategoryDTO;
-import shpp.azaika.dto.StoreDTO;
 
 import java.sql.*;
 import java.util.ArrayList;
@@ -18,7 +17,7 @@ public class CategoryDAO {
         this.connection = connection;
     }
 
-    public List<Long> insertBatch(List<CategoryDTO> dtos) {
+    public List<Short> insertBatch(List<CategoryDTO> dtos) {
         String sql = "INSERT INTO categories (name) VALUES (?)";
         try (PreparedStatement stmt = connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)){
             for (CategoryDTO dto : dtos) {
@@ -33,9 +32,9 @@ public class CategoryDAO {
         }
     }
 
-    public List<Long> insertInChunks(List<CategoryDTO> dtos, int chunkSize) {
+    public List<Short> insertInChunks(List<CategoryDTO> dtos, int chunkSize) {
         int total = dtos.size();
-        List<Long> ids = new ArrayList<>();
+        List<Short> ids = new ArrayList<>();
         for (int i = 0; i < total; i += chunkSize) {
             int end = Math.min(i + chunkSize, total);
             List<CategoryDTO> chunk = dtos.subList(i, end);
@@ -44,10 +43,10 @@ public class CategoryDAO {
         return ids;
     }
 
-    private List<Long> retrieveIds(ResultSet generatedKeys) throws SQLException {
-        List<Long> ids = new ArrayList<>();
+    private List<Short> retrieveIds(ResultSet generatedKeys) throws SQLException {
+        List<Short> ids = new ArrayList<>();
         while (generatedKeys.next()){
-            ids.add(generatedKeys.getLong(1));
+            ids.add(generatedKeys.getShort(1));
         }
         return ids;
     }
